Conflicts:
	Projet_service_client/src/fr/dauphine/main/Connexion.java
	projet_service_web/src/fr/dauphine/beans/BibliothequeImpl.java
	projet_service_web/src/fr/dauphine/interfaces/Bibliotheque.java

+++ Projet_service_client/src/fr/dauphine/widgets/SelectionListener.java
@@ -1,6 +1,6 @@
 package fr.dauphine.widgets;
 
 public interface SelectionListener {
-	void selectionLigne(Integer id);
+	void selectionLigne(long id);
 
 }
